backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-dotenv==1.0.0
openai>=1.12.0
httpx>=0.25.0
tavily-python==0.3.0
alembic>=1.13.0
pytest>=7.4.0
pytest-asyncio>=0.21.0



